feat: Groupエンティティ作成
